8.13. Пришёл, увидел, загрузил (часть 1) (fix 2)
