+++ src/utils/point.js
@@ -22,3 +22,23 @@ export const isDatesEqual = (dateA, dateB) => {
 export const isPointExpired = (date) => {
   return date === null ? false : dayjs().isAfter(date, 'D');
 };
+
+export const generateTripInfo = (points) => {
+  if (points.length === 0 ) return {};
+  const title = (points.length <= 3) ?
+    points.map((point)=> point.destination.name).join('-') :
+    points[0].destination.name + '- ... -' + points[points.length - 1].destination.name;
+
+
+  const price = points.reduce((sum , point) => {
+    return sum  + Number(point.price);
+  }, 0);
+
+  const date =  dayjs(points[0].startDate).format('DD MMM') +'-' +  dayjs(points[points.length - 1].endDate).format('DD MMM');
+
+  return {
+    title,
+    date,
+    price,
+  };
+};
